Restore ability to create SRPMs
rpmbuild (the github workflow tool) is getting further along in its job
but is failing when running `rpmbuild -bb` in the container. Instead
of trying to figure out why that's busted, maybe I should just
restore SRPM creation first. That's what this commit does.

Previously, the project created an SRPM by hitting github to get
a tarball of the source.

Now, we'll just use the existing `/github/workspace/` in the
Docker container. This includes the `.git/` subdir of the
project, which we take extra steps to clean up.