Make rpmbuild work with private repos
Previously, in order to create /github/home/rpmbuild/SOURCES/foo-1.2.0.tar.gz,
rpmbuild did the following:

1. use curl to download a tarball of the project source
2. unpack it into a local directory
3. repack it with the correct desired directory structure
4. move the repacked tarball to /github/home/rpmbuild/SOURCES

This failed for me at step 1 because the repo I'm trying to use rpmbuild on
is a private repo. There's no means of plumbing a github auth token into the
`curl` command, so the tarball fails to download and everything else fails.

----

Since we have the current repo tree in /github/workspace, we can coerce `git`
to make the archive for us. Now, we:

1. ask git to make the archive

and everything works with a private repo. In order to make this approach work,
I've added `git` into the build container.